Update 14
Changed the dashes with underscores in the Content file, while also replacing them with the Monogame Pipeline Tool. Still, there seem to be some errors regarding the name change of the classes. I wonder what to do. Should we revert the names back? Because I don't know how to handle this problem.